maint: switch to using pyproject.toml instead of setup.py/cfg
