--- MANIFEST.in
@@ -1,12 +1,0 @@
-include *.md
-include LICENSE
-include hera_qm/VERSION
-include hera_qm/GIT_INFO
-include *.bib
-include scripts/*.pro
-include scripts/*.sh
-recursive-include hera_qm/data *
-prune hera_qm/data/test
-recursive-include docs *
-exclude docs/.DS_Store
-prune docs/_build
