+++ hera_qm/scripts/__init__.py
@@ -1,0 +1,1 @@
+"""Dummy package for hera_qm scripts."""
