+++ hera_qm/scripts/test_xrfi_performance.py
