--- setup.cfg
@@ -1,13 +1,0 @@
-[aliases]
-test=pytest
-
-[tool:pytest]
-addopts = --ignore=scripts --cov=hera_qm --cov-config=.coveragerc
-
-[pycodestyle]
-ignore = E501,W503
-filename = *.py
-exclude = build/,docs/,.pytest_cache,.svn,CVS,.bzr,.hg,.git,__pycache__,dist/,scripts/,hera_qm.egg-info,hera_qm/data/
-
-[pydocstyle]
-convention = numpy
